python/packages/jumpstarter-driver-nanokvm/pyproject.toml
[project]
name = "jumpstarter-driver-nanokvm"
dynamic = ["version", "urls"]
description = "Jumpstarter driver for NanoKVM devices providing video streaming, keyboard/mouse control, and serial console access"
readme = "README.md"
license = "Apache-2.0"
authors = [
    { name = "Miguel Angel Ajo", email = "miguelangel@ajo.es" }
]
requires-python = ">=3.11"
dependencies = [
    "anyio>=4.10.0",
    "jumpstarter",
    "jumpstarter-driver-composite",
    "jumpstarter-driver-pyserial",
    "nanokvm @ git+https://github.com/mangelajo/python-nanokvm.git@dev",
    "aiohttp",
    "pillow",
    "pydantic",
    "yarl",
    "click",
]

[project.entry-points."jumpstarter.drivers"]
NanoKVM = "jumpstarter_driver_nanokvm.driver:NanoKVM"

[tool.hatch.version]
source = "vcs"
raw-options = { 'root' = '../../..'}

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.metadata.hooks.vcs.urls]
Homepage = "https://jumpstarter.dev"
source_archive = "https://github.com/jumpstarter-dev/repo/archive/{commit_hash}.zip"

[tool.pytest.ini_options]
addopts = "--cov --cov-report=html --cov-report=xml"
log_cli = true
log_cli_level = "INFO"
testpaths = ["jumpstarter_driver_nanokvm"]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling", "hatch-vcs", "hatch-pin-jumpstarter"]
build-backend = "hatchling.build"

[tool.hatch.build.hooks.pin_jumpstarter]
name = "pin_jumpstarter"

[dependency-groups]
dev = [
    "pytest-cov>=6.0.0",
    "pytest>=8.3.3",
]
